Refactor SSM runbook link action: modularize properties into reusable components (`runbookNameProperty`, `runbookVersionProperty`, `runbookParametersProperty`) and extract AWS operations helpers to improve readability and maintainability.
