+++ packages/openops/src/index.ts
@@ -22,8 +22,12 @@ export * from './lib/aws/rds/rds-delete-instance';
 export * from './lib/aws/rds/rds-delete-snapshot';
 export * from './lib/aws/rds/rds-describe';
 export * from './lib/aws/regions';
+export * from './lib/aws/ssm/get-ssm-describe-document-info';
 export * from './lib/aws/ssm/get-ssm-document-versions';
 export * from './lib/aws/ssm/get-ssm-documents';
+export * from './lib/aws/ssm/runbook-name-property';
+export * from './lib/aws/ssm/runbook-parameters-property';
+export * from './lib/aws/ssm/runbook-version-property';
 export * from './lib/aws/sts-common';
 export * from './lib/aws/tags/tag-resources';
 export * from './lib/aws/tags/tags';
